+++ src/datarefinery/plugins/image_classification/__init__.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2026 Pointmatic
+# SPDX-License-Identifier: Apache-2.0
+
+from datarefinery.plugins.image_classification.plugin import (
+    PLUGIN,
+    ImageClassificationPlugin,
+)
+
+__all__ = ["PLUGIN", "ImageClassificationPlugin"]